+++ CHANGELOG.md
@@ -12,6 +12,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.4.4] - 2021-04-05
+
+### Added
+
++ tests
+
 ## [0.4.3] - 2021-04-05
 
 ### Added